requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.5
requests

google-api-python-client
google-auth
google-auth-oauthlib
apache-airflow[google]
google-cloud
google-cloud-bigquery
google-cloud-storage

# google-api-python-client==1.7.2
# google-auth==1.8.0
# google-auth-httplib2==0.0.3
# google-auth-oauthlib==0.4.1